test: preserve DankEth middleware behavior
Extract shared JSON-RPC unit-test helpers so follow-up fixes stay focused.

Add a guardrail proving DankEth still honors non-Dank async middleware.
